+++ gttk/utils/exceptions.py
@@ -18,14 +18,6 @@ class ProcessingStepFailedError(RuntimeError):
     """Custom exception raised for errors during a GeoTIFF processing step."""
     pass
 
-class TransparencyProcessingError(Exception):
-    """Base exception for errors during transparency processing."""
-    pass
-
-class ValidateCOGError(Exception):
-    """Custom exception for errors during COG validation."""
-    pass
-
 class GdalExecutionError(RuntimeError):
     """Custom exception for errors during GDAL command execution."""
     pass